# openapi.yaml
openapi: 3.0.3
info:
  title: Skills-FS
  version: '1.0'
  description: Admin API for managing Agent Skills and mount tokens.
servers:
  - url: /api/v1

paths:
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      description: List all skills, newest metadata only (no bodies).
      responses:
        '200':
          description: List of skills.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SkillSummary' }
    post:
      operationId: createSkill
      tags: [skills]
      description: Create a skill.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkillWrite' }
      responses:
        '201':
          description: Created skill.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Skill' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /skills/{id}:
    parameters:
      - $ref: '#/components/parameters/SkillID'
    get:
      operationId: getSkill
      tags: [skills]
      description: Get a single skill including its body.
      responses:
        '200':
          description: Skill.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Skill' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateSkill
      tags: [skills]
      description: Replace all editable fields of a skill.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkillWrite' }
      responses:
        '200':
          description: Updated skill.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Skill' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: deleteSkill
      tags: [skills]
      description: Delete a skill.
      responses:
        '204':
          description: Deleted.
        '404': { $ref: '#/components/responses/NotFound' }
  /config:
    get:
      operationId: getConfig
      tags: [meta]
      description: Non-secret server configuration, e.g. whether skill generation is enabled.
      responses:
        '200':
          description: Server configuration.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerConfig' }
  /generate:
    post:
      operationId: createGeneration
      tags: [generate]
      description: >-
        Start asynchronous skill generation from a raw idea. The job runs in the
        background; poll GET /generate/{id} for the result. Disabled unless the
        LLM API key is configured at startup.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerationWrite' }
      responses:
        '202':
          description: Generation job accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /generate/{id}:
    parameters:
      - $ref: '#/components/parameters/GenerationID'
    get:
      operationId: getGeneration
      tags: [generate]
      description: Get the state of a generation job.
      responses:
        '200':
          description: Job state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Generation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      description: List all mount tokens (no secrets).
      responses:
        '200':
          description: List of tokens.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Token' }
    post:
      operationId: createToken
      description: Create a mount token. The plaintext token is returned once.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenWrite' }
      responses:
        '201':
          description: Created token with plaintext secret.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /tokens/{id}:
    parameters:
      - $ref: '#/components/parameters/TokenID'
    delete:
      operationId: deleteToken
      tags: [tokens]
      description: Revoke a mount token.
      responses:
        '204':
          description: Deleted.
        '404': { $ref: '#/components/responses/NotFound' }

components:
  parameters:
    SkillID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    TokenID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    GenerationID:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict with existing resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Metadata:
      type: object
      additionalProperties: { type: string }
      description: Arbitrary string key-value metadata (SKILL.md frontmatter).
    SkillWrite:
      type: object
      description: Editable skill fields. Optional fields may be omitted.
      required: [name, description, body]
      properties:
        name:
          type: string
          description: Skill name = directory name. Lowercase a-z0-9 and hyphens, max 64 chars.
        description:
          type: string
          maxLength: 1024
        body:
          type: string
          description: SKILL.md markdown body after frontmatter.
        license:
          type: string
        compatibility:
          type: string
          maxLength: 500
        allowedTools:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
    SkillSummary:
      type: object
      required: [id, name, description, license, compatibility, allowedTools, metadata, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        license: { type: string }
        compatibility: { type: string }
        allowedTools: { type: string }
        metadata: { $ref: '#/components/schemas/Metadata' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Skill:
      allOf:
        - $ref: '#/components/schemas/SkillSummary'
        - type: object
          required: [body]
          properties:
            body: { type: string }
    Token:
      type: object
      required: [id, label, prefix, lastUsedAt, createdAt]
      properties:
        id: { type: integer, format: int64 }
        label: { type: string }
        prefix: { type: string, description: First characters of the token for identification. }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    TokenCreated:
      allOf:
        - $ref: '#/components/schemas/Token'
        - type: object
          required: [token]
          properties:
            token: { type: string, description: Plaintext token; shown only at creation. }
    TokenWrite:
      type: object
      required: [label]
      properties:
        label: { type: string }
    ServerConfig:
      type: object
      required: [llm]
      properties:
        llm:
          type: object
          required: [enabled, model]
          properties:
            enabled:
              type: boolean
              description: Whether agent-based skill generation is configured (LLM API key set at startup).
            model:
              type: string
              description: Model used for generation.
    GenerationWrite:
      type: object
      required: [idea]
      properties:
        idea:
          type: string
          minLength: 1
          maxLength: 20000
          description: Raw idea or draft to turn into a skill.
    GenerationCreated:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: Generation job id, usable with GET /generate/{id}.
    Generation:
      type: object
      required: [id, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [running, done, error]
        error:
          type: string
          description: Human-readable failure reason when status is error.
        skillId:
          type: integer
          format: int64
          description: Created skill id when status is done.
        skillName:
          type: string
          description: Created skill name when status is done.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }